+++ source/gnss/gnssParametrization/gnssParametrizationIonosphereMap.h
@@ -18,19 +18,25 @@
 #ifdef DOCSTRING_GnssParametrization
 static const char *docstringGnssParametrizationIonosphereMap = R"(
 \subsection{IonosphereMap}\label{gnssParametrizationType:ionosphereMap}
-TODO: reading and writing ionosphere maps not implemented yet.
-% A priori ionopshere model can be provided with \configFile{inputfileMap}{gnssIonosphereMaps}.
+Apriori VTEC maps can be removed from the observations with
+\configFile{inputfileGriddedDataTimeSeries}{griddedDataTimeSeries}
+(e.g. from \program{GnssIonex2GriddedDataTimeSeries}).
 
-The ionosphere is parametrized in terms of $[TECU]$ as a \configClass{temporal}{parametrizationTemporalType}ly
+The ionosphere is parametrized in terms of $[TECU]$ in a single layer sphere with
+\config{radiusIonosphericLayer} as a \configClass{temporal}{parametrizationTemporalType}ly
 changing (e.g. hourly linear splines) spherical harmonics expansion
 \begin{equation}
   VTEC(\lambda,\theta,t) = \sum_{n=0}^{n_{max}} \sum_{m=0}^n c_{nm}(t)C_{nm}(\lambda,\theta)+s_{nm}(t)S_{nm}(\lambda,\theta)
 \end{equation}
 up to \config{maxDegree}=\verb|15| in a solar-geomagentic frame defined
-by \configClass{magnetosphere}{magnetosphereType}.
-The VTEC values are mapped to STEC values via eq.~\eqref{gnssParametrizationType:IonosphereVTEC:STEC}.
+by \configClass{magnetosphere}{magnetosphereType}. The VTEC values are mapped to STEC values
+in the observation equations via eq.~\eqref{gnssParametrizationType:IonosphereVTEC:STEC}.
 
-Local and short-term scintillations can be considered by adding constrained
+The estimated VTEC inclusive the apriori \configFile{inputfileGriddedDataTimeSeries}{griddedDataTimeSeries}
+can be written to \configFile{outputfileGriddedDataTimeSeries}{griddedDataTimeSeries}
+evaluated at \configClass{outputGrid}{gridType} and \configClass{outputTimeSeries}{timeSeriesType}.
+
+Local and short-term scintillations should be considered by adding constrained
 \configClass{parametrization:ionosphereSTEC}{gnssParametrizationType:ionosphereSTEC}.
 To account for signal biases add
 \configClass{parametrization:tecBiases}{gnssParametrizationType:tecBiases}.
@@ -46,6 +52,7 @@ The \file{parameter names}{parameterName} are
 /***********************************************/
 
 #include "base/import.h"
+#include "base/griddedData.h"
 #include "config/config.h"
 #include "classes/magnetosphere/magnetosphere.h"
 #include "classes/parametrizationTemporal/parametrizationTemporal.h"
@@ -65,15 +72,18 @@ class GnssParametrizationIonosphereMap : public GnssParametrizationBase
   PlatformSelectorPtr             selectReceivers;
   std::vector<Byte>               selectedReceivers;
   FileName                        fileNameIn, fileNameOut;
+  GriddedData                     gridOut;
+  std::vector<Time>               timesOut;
   UInt                            maxDegree;
   ParametrizationTemporalPtr      temporal;
-  Double                          mapR, mapH, mapAlpha;
+  Double                          radiusIono, mapR, mapH, mapAlpha;
   MagnetospherePtr                magnetosphere;
   std::vector<Vector>             x;
   std::vector<GnssParameterIndex> index;
 
-  Vector3d intersection(const Vector3d &posRecv, const Vector3d &posTrans, Angle elevation) const;
+  Vector3d intersection(const Double radiusIono, const Vector3d &posRecv, const Vector3d &posTrans) const;
   Double   mapping(Angle elevation) const;
+  Double   sphericalHarmonicSynthesis(const Vector3d &point, const Vector &x) const;
 
 public:
   GnssParametrizationIonosphereMap(Config &config);